ai-services/requirements.txt
# Core AI and Machine Learning libraries
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0

# Computer Vision and Facial Analysis
dlib>=19.24.0
face-recognition>=1.3.0
mediapipe>=0.10.5
imutils>=0.5.4

# Audio Processing for Voice Analysis
librosa>=0.10.0
pyaudio>=0.2.11
soundfile>=0.12.0
pydub>=0.25.0
speechrecognition>=3.10.0

# Google Gemini API
google-generativeai>=0.3.0

# Web framework for API endpoints
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.3.0
websockets>=11.0.0
wsproto>=1.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
